# src/main/java/com/inventorilab/response/PagingResponse.java
package com.inventorilab.response;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Builder
@AllArgsConstructor
@NoArgsConstructor
public class PagingResponse {
    private int currentPage;
    private int totalPages;
    private int size;
    private long totalElements;
}